// src/model/UserData.ts
export interface UserData {
  firstname: string;
  lastname: string;
  mail: string;
  slogan: string;
  job: string;
  company: string;
  //experience: any[];
  topskill1: string;
  topskill2: string;
  topskill3: string;
  // phone: any;
  //street: any;
  //street_number: any;
  //place: any;
  //design_card: any;
  //Html: any;
  location: string;
}